# Option for benchmarking with user defined data