This adds the option to benchmark with user defined data. It is
specifically use full for benchmarking file systems with compression, as
IOR only provides in-compressible or super compressible data (repeating
every 128 bytes). A new flag for the -l (data packet type) `u` is now
present. When this is chosen the data given by the new directive
`pathToInputFile` is read to the buffer. The file has to be at least
the size of blocksize as this is amount of user data used by IOR.
This seams to be sufficient for simulating real data while keeping IOR's
options for access patterns through the parameters transfersize,
blocksize, segmentcount and number of processes working.
This option shouldn't alter IOR's behavior in any other way.
User data should be compatible with Read and Write Checking as the
buffers can be filled with the same data for checking.