@ezickler 2018-01-11
Sorry messed up my branch while fixing a bug and #23 got closed.