@johnbent 2018-01-11
If the file is just block-size, then every block will be compressible and
identical.  I wonder if that is good enough?  Would it be possible to have
a unique file per process?  And the file be the total size of the data
written?

On Thu, Jan 11, 2018 at 9:46 AM, Enno Zickler <notifications@github.com>
wrote:

> Sorry messed up my branch while fixing a bug and #23
> <https://github.com/IOR-LANL/ior/pull/23> got closed.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/IOR-LANL/ior/pull/27#issuecomment-356987808>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AB89PjBUFLVKqm8HqGhHIcDOae00jbN1ks5tJjrbgaJpZM4RbHfG>
> .
>