@glennklockwood 2018-01-11
Thanks for contributing this, @ezickler!

I had originally thought that each rank would read in from a different offset in the input file, then rewrite that unique block out to get around this.  As implemented now though, I would guess that all you have to do is know the underlying compression algorithm's block size and make sure that your IOR blocks are larger than that to work around the compressibility of the periodic data.

In the interests of keeping the test procedure for accepting this manageable though, I'd prefer to verify this implementation before we move on to doing something different for each rank.  Verifying the current implementation might be a fair amount of work because this patch touches the critical data path, though.  I would like to

1. ensure that the pre- and post-patch version of the code produces the same bandwidth measurements on a given file system (I am confident that it will)

2. ensure that the pre- and post-patch version of the code produces the same IOPs measurements (4k random read/write) on flash or ramdisk.  I'm a little concerned that the extra logic around the larger blockSize-sized buffer might affect IOR's latency measurements.  It shouldn't, but I'd feel better if we could prove it.

I personally won't be able to review this code for a month or two at least, so if anyone else wants to make sure it looks good, please go ahead.