@ezickler 2018-01-14
Thanks to both of you for your input.

The reason for choosing blocksize as the amount of data for the input is that it seams to be easy enough to keep the behavior of ior pretty close to the original but also to make it reasonably big an then chunk it with transfersize to small accesses.  
I thought of an independent version, where an arbitrary long input file could be loaded to a ring buffer an this would be accessed continuously. I still like the idea but i think it make the output much less easy to predict and verify manually. Also i'm not sure how good  a ring buffer would be performance wise. And page alignment wouldn't work any more.
Another reason for the buffer sized input is that it allows to have an know data set of e.g 1GB as input and run a benchmark with lots of clients and bench a X TB file. And i don't like the idea of pregenerating a X TB file for benchmarking.  This is by the way the use case i'm working towards. 
I like the idea with different data for every rank, but i think if the input amount of data is limited in relation to the output, with IOR's randomization options similar diversity off accesses for io clients to servers are generated. But haven't thought it through in detail yet.

About making the blocks only as big as the underlying compression algorithm, i don't think that would be enough. Aren't there algorithms which are checking for more than just a small window.  More like deduplicating file systems. And more important this would mean one could only check the performance for the compressiblety of this on block which stays the same but is sill far for real world data.

On the verification i'm totally with you. This new feature shouldn't alter the behavior of the old functionality. I will look into your proposed tests.  Maybe i can do some verification. 
Any detailed configs this test should be done on?
With 4k random read/write, did you mean 4k IOR blocks or more like storage benchmarks 4k accesses which would mean 4k transfers i think.
The randomization from IOR should do the random part.
But as you suggest a ram disk, a local test is sufficient or should it be done on a cluster with lustre and SSDs? Doesn't seams right for me as we wonna  test if the IOPs IOR can produce are limited. Am i right?