@ezickler 2018-02-19
So i did some benchmarking today. It was done on a system equipped with 4 Opteron 6168 and 128GiB of DDR3 memory. A single memory channel could handle up to 1600 MiB/s . As the system has 4 CPU of which each has four channels this system can handle up to 25600 MiB/s if all channels are used. A 100GiB tmpfs was used for storage. 

I did both of the proposed benchmarks once single threaded and also with 48 processes. `filePerProc` so writing with 48 proc can also use all 16 memory channels. `reorderTasksRandom` and `fsync` to prevent caching effects. 

I'm not quite sure what these results means, as the `userDataOption` branch is often faster. Especially the `4k random 1proc` case where it's close to a 10% increase while the standard deviation is small. Maybe i just did to few iterations or there is an error in my new implementation. At least there seams nothing decreasing the performance.  What do you think? 

Gathered results:

|                 | write - master | write - userDataOption |  read - master  | read - userDataOption |
| ------------ |:------------: | :-------------: | :-------------: | :-------------: |
  **4k random 1 proc**    |  916.09 | 999.34 | 1095.89 | 1415.24
  **4k random 48 proc**  | 24363.48  |  25077.63 | 22211.49 | 22423.20
  **512m seq 1proc**      |  1207.94 | 1169.0 | 1619.55 |  1556.16
  **512m seq 48 proc**   |   22253.88 |  22383.37 | 17372.22 | 18151.47

IOR script:
```
IOR START
	testFile = /mnt/tmpfs/ior-bench.data
	api=POSIX
	filePerProc=1
	randomOffset=1
	fsync=1
	repetitions=5
    reordertasksrandom=1
	transferSize=4k
	blockSize=96g
	numTasks=1
	RUN
	blockSize=2g
	numTasks=48
	RUN
	randomOffset=0
	transferSize=512m
	blockSize=2g
	numTasks=48
	RUN
	blockSize=96g
	numTasks=1
	RUN
IOR STOP
```
Test summerys:
```
Current master:
Summary of all tests:
Operation   Max(MiB)   Min(MiB)  Mean(MiB)     StdDev    Mean(s) Test# #Tasks tPN reps fPP reord reordoff reordrand seed segcnt blksiz         xsize     aggsize      API   RefNum
write         919.72     912.80     916.09       2.59  107.30915 0     1      1   5    1   0     1        1         0    1      103079215104   4096      103079215104 POSIX 0
read         1096.64    1095.02    1095.89       0.53   89.70208 0     1      1   5    1   0     1        1         0    1      103079215104   4096      103079215104 POSIX 0
write       24744.72   23113.73   24363.48     626.31    4.03767 1     48     48  5    1   0     1        1         0    1      2147483648     4096      103079215104 POSIX 0
read        23959.80   21012.30   22211.49    1419.94    4.44352 1     48     48  5    1   0     1        1         0    1      2147483648     4096      103079215104 POSIX 0
write       22528.49   21218.61   22253.88     517.78    4.41987 2     48     48  5    1   0     1        1         0    1      2147483648     536870912 103079215104 POSIX 0
read        17674.00   17114.98   17372.22     177.73    5.65928 2     48     48  5    1   0     1        1         0    1      2147483648     536870912 103079215104 POSIX 0
write        1236.11    1157.48    1207.94      26.77   81.42220 3     1      1   5    1   0     1        1         0    1      103079215104   536870912 103079215104 POSIX 0
read         1652.64    1524.03    1619.55      48.17   60.75465 3     1      1   5    1   0     1        1         0    1      103079215104   536870912 103079215104 POSIX 0


With new Option:
Summary of all tests:
Operation   Max(MiB)   Min(MiB)  Mean(MiB)     StdDev    Mean(s) Test# #Tasks tPN reps fPP reord reordoff reordrand seed segcnt blksiz          xsize    aggsize      API   RefNum
write        1002.55     995.50     999.34       2.65   98.37003 0     1      1   5    1   0     1        1         0    1      103079215104   4096      103079215104 POSIX 0
read         1417.42    1408.93    1415.24       3.18   69.46142 0     1      1   5    1   0     1        1         0    1      103079215104   4096      103079215104 POSIX 0
write       25154.87   25012.36   25077.63      49.63    3.92000 1     48     48  5    1   0     1        1         0    1      2147483648     4096      103079215104 POSIX 0
read        22659.51   22329.83   22423.20     119.66    4.38416 1     48     48  5    1   0     1        1         0    1      2147483648     4096      103079215104 POSIX 0
write       22391.01   22375.47   22383.37       5.94    4.39183 2     48     48  5    1   0     1        1         0    1      2147483648     536870912 103079215104 POSIX 0
read        18205.17   18058.65   18151.47      51.52    5.41580 2     48     48  5    1   0     1        1         0    1      2147483648     536870912 103079215104 POSIX 0
write        1207.75    1129.12    1169.08      31.33   84.14705 3     1      1   5    1   0     1        1         0    1      103079215104   536870912 103079215104 POSIX 0
read         1617.15    1483.18    1556.16      49.67   63.23555 3     1      1   5    1   0     1        1         0    1      103079215104   536870912 103079215104 POSIX 0
```

Complete output:
```
IOR-3.1.0: MPI Coordinated Test of Parallel I/O

Began: Mon Feb 19 13:24:28 2018
Command line used: ./ior-master -f userDataBench.ior
Machine: Linux magny1

Test 0 started: Mon Feb 19 13:24:28 2018
Summary:
	api                = POSIX
	test filename      = /mnt/tmpfs/ior-bench.data
	access             = file-per-process
	ordering in a file = random offsets
	ordering inter file= random task offsets >= 1, seed=0
	clients            = 1 (1 per node)
	repetitions        = 5
	xfersize           = 4096 bytes
	blocksize          = 96 GiB
	aggregate filesize = 96 GiB

access    bw(MiB/s)  block(KiB) xfer(KiB)  open(s)    wr/rd(s)   close(s)   total(s)   iter
------    ---------  ---------- ---------  --------   --------   --------   --------   ----
write     913.54     100663296  4.00       0.000015   107.61     0.000016   107.61     0   
read      1096.16    100663296  4.00       0.000017   89.68      0.000015   89.68      0   
remove    -          -          -          -          -          -          35.59      0   
write     917.74     100663296  4.00       0.000033   107.12     0.000014   107.12     1   
read      1095.77    100663296  4.00       0.000018   89.71      0.000014   89.71      1   
remove    -          -          -          -          -          -          34.80      1   
write     912.80     100663296  4.00       0.000033   107.70     0.000014   107.70     2   
read      1095.02    100663296  4.00       0.000017   89.77      0.000013   89.77      2   
remove    -          -          -          -          -          -          35.29      2   
write     919.72     100663296  4.00       0.000033   106.88     0.000014   106.88     3   
read      1096.64    100663296  4.00       0.000017   89.64      0.000014   89.64      3   
remove    -          -          -          -          -          -          35.16      3   
write     916.65     100663296  4.00       0.000033   107.24     0.000014   107.24     4   
read      1095.89    100663296  4.00       0.000016   89.70      0.000014   89.70      4   
remove    -          -          -          -          -          -          34.72      4   

Max Write: 919.72 MiB/sec (964.40 MB/sec)
Max Read:  1096.64 MiB/sec (1149.91 MB/sec)

Test 1 started: Mon Feb 19 13:43:48 2018
Summary:
	api                = POSIX
	test filename      = /mnt/tmpfs/ior-bench.data
	access             = file-per-process
	ordering in a file = random offsets
	ordering inter file= random task offsets >= 1, seed=0
	clients            = 48 (48 per node)
	repetitions        = 5
	xfersize           = 4096 bytes
	blocksize          = 2 GiB
	aggregate filesize = 96 GiB

access    bw(MiB/s)  block(KiB) xfer(KiB)  open(s)    wr/rd(s)   close(s)   total(s)   iter
------    ---------  ---------- ---------  --------   --------   --------   --------   ----
write     24677      2097152    4.00       0.012558   3.98       0.479929   3.98       0   
read      21076      2097152    4.00       0.011801   4.66       3.04       4.66       0   
remove    -          -          -          -          -          -          0.827445   0   
write     24745      2097152    4.00       0.012605   3.97       0.529506   3.97       1   
read      21068      2097152    4.00       0.011570   4.67       3.05       4.67       1   
remove    -          -          -          -          -          -          0.835183   1   
write     24671      2097152    4.00       0.012404   3.98       0.476646   3.98       2   
read      21012      2097152    4.00       0.011847   4.68       3.07       4.68       2   
remove    -          -          -          -          -          -          0.834541   2   
write     23114      2097152    4.00       0.023549   4.25       0.782789   4.25       3   
read      23960      2097152    4.00       0.012099   4.10       2.50       4.10       3   
remove    -          -          -          -          -          -          1.09       3   
write     24611      2097152    4.00       0.012515   3.99       0.516242   3.99       4   
read      23941      2097152    4.00       0.011850   4.11       2.51       4.11       4   
remove    -          -          -          -          -          -          0.937730   4   

Max Write: 24744.72 MiB/sec (25946.72 MB/sec)
Max Read:  23959.80 MiB/sec (25123.67 MB/sec)

Test 2 started: Mon Feb 19 13:44:35 2018
Summary:
	api                = POSIX
	test filename      = /mnt/tmpfs/ior-bench.data
	access             = file-per-process
	ordering in a file = sequential offsets
	ordering inter file= random task offsets >= 1, seed=0
	clients            = 48 (48 per node)
	repetitions        = 5
	xfersize           = 512 MiB
	blocksize          = 2 GiB
	aggregate filesize = 96 GiB

access    bw(MiB/s)  block(KiB) xfer(KiB)  open(s)    wr/rd(s)   close(s)   total(s)   iter
------    ---------  ---------- ---------  --------   --------   --------   --------   ----
write     21219      2097152    524288     0.012816   4.63       0.875134   4.63       0   
read      17115      2097152    524288     0.011816   5.74       2.45       5.74       0   
remove    -          -          -          -          -          -          0.387489   0   
write     22528      2097152    524288     0.012843   4.36       0.592319   4.36       1   
read      17354      2097152    524288     0.011946   5.66       2.51       5.66       1   
remove    -          -          -          -          -          -          0.362151   1   
write     22494      2097152    524288     0.012837   4.37       0.617450   4.37       2   
read      17362      2097152    524288     0.011295   5.66       2.51       5.66       2   
remove    -          -          -          -          -          -          0.360260   2   
write     22522      2097152    524288     0.012947   4.36       0.598568   4.36       3   
read      17674      2097152    524288     0.012295   5.56       2.56       5.56       3   
remove    -          -          -          -          -          -          0.363252   3   
write     22507      2097152    524288     0.013039   4.37       0.614290   4.37       4   
read      17356      2097152    524288     0.012303   5.66       2.65       5.66       4   
remove    -          -          -          -          -          -          0.363468   4   

Max Write: 22528.49 MiB/sec (23622.84 MB/sec)
Max Read:  17674.00 MiB/sec (18532.53 MB/sec)

Test 3 started: Mon Feb 19 13:45:31 2018
Summary:
	api                = POSIX
	test filename      = /mnt/tmpfs/ior-bench.data
	access             = file-per-process
	ordering in a file = sequential offsets
	ordering inter file= random task offsets >= 1, seed=0
	clients            = 1 (1 per node)
	repetitions        = 5
	xfersize           = 512 MiB
	blocksize          = 96 GiB
	aggregate filesize = 96 GiB

access    bw(MiB/s)  block(KiB) xfer(KiB)  open(s)    wr/rd(s)   close(s)   total(s)   iter
------    ---------  ---------- ---------  --------   --------   --------   --------   ----
write     1218.91    100663296  524288     0.000045   80.65      0.000013   80.65      0   
read      1635.24    100663296  524288     0.000017   60.12      0.000015   60.12      0   
remove    -          -          -          -          -          -          12.15      0   
write     1157.48    100663296  524288     0.000045   84.93      0.000014   84.93      1   
read      1524.03    100663296  524288     0.000017   64.50      0.000015   64.50      1   
remove    -          -          -          -          -          -          12.12      1   
write     1218.98    100663296  524288     0.000042   80.64      0.000013   80.64      2   
read      1647.71    100663296  524288     0.000017   59.66      0.000016   59.66      2   
remove    -          -          -          -          -          -          12.10      2   
write     1236.11    100663296  524288     0.000042   79.53      0.000015   79.53      3   
read      1638.10    100663296  524288     0.000017   60.01      0.000015   60.01      3   
remove    -          -          -          -          -          -          12.12      3   
write     1208.24    100663296  524288     0.000044   81.36      0.000013   81.36      4   
read      1652.64    100663296  524288     0.000017   59.48      0.000015   59.48      4   
remove    -          -          -          -          -          -          12.07      4   

Max Write: 1236.11 MiB/sec (1296.16 MB/sec)
Max Read:  1652.64 MiB/sec (1732.92 MB/sec)

Summary of all tests:
Operation   Max(MiB)   Min(MiB)  Mean(MiB)     StdDev    Mean(s) Test# #Tasks tPN reps fPP reord reordoff reordrand seed segcnt blksiz xsize aggsize API RefNum
write         919.72     912.80     916.09       2.59  107.30915 0 1 1 5 1 0 1 1 0 1 103079215104 4096 103079215104 POSIX 0
read         1096.64    1095.02    1095.89       0.53   89.70208 0 1 1 5 1 0 1 1 0 1 103079215104 4096 103079215104 POSIX 0
write       24744.72   23113.73   24363.48     626.31    4.03767 1 48 48 5 1 0 1 1 0 1 2147483648 4096 103079215104 POSIX 0
read        23959.80   21012.30   22211.49    1419.94    4.44352 1 48 48 5 1 0 1 1 0 1 2147483648 4096 103079215104 POSIX 0
write       22528.49   21218.61   22253.88     517.78    4.41987 2 48 48 5 1 0 1 1 0 1 2147483648 536870912 103079215104 POSIX 0
read        17674.00   17114.98   17372.22     177.73    5.65928 2 48 48 5 1 0 1 1 0 1 2147483648 536870912 103079215104 POSIX 0
write        1236.11    1157.48    1207.94      26.77   81.42220 3 1 1 5 1 0 1 1 0 1 103079215104 536870912 103079215104 POSIX 0
read         1652.64    1524.03    1619.55      48.17   60.75465 3 1 1 5 1 0 1 1 0 1 103079215104 536870912 103079215104 POSIX 0

Finished: Mon Feb 19 13:58:23 2018
IOR-3.1.0: MPI Coordinated Test of Parallel I/O

Began: Mon Feb 19 13:58:24 2018
Command line used: ./ior-userdata -f userDataBench.ior
Machine: Linux magny1

Test 0 started: Mon Feb 19 13:58:24 2018
Summary:
	api                = POSIX
	test filename      = /mnt/tmpfs/ior-bench.data
	access             = file-per-process
	ordering in a file = random offsets
	ordering inter file= random task offsets >= 1, seed=0
	clients            = 1 (1 per node)
	repetitions        = 5
	xfersize           = 4096 bytes
	blocksize          = 96 GiB
	aggregate filesize = 96 GiB

access    bw(MiB/s)  block(KiB) xfer(KiB)  open(s)    wr/rd(s)   close(s)   total(s)   iter
------    ---------  ---------- ---------  --------   --------   --------   --------   ----
write     997.49     100663296  4.00       0.000021   98.55      0.000014   98.55      0   
read      1417.42    100663296  4.00       0.000014   69.35      0.000012   69.35      0   
remove    -          -          -          -          -          -          12.10      0   
write     1001.89    100663296  4.00       0.000027   98.12      0.000012   98.12      1   
read      1416.77    100663296  4.00       0.000013   69.39      0.000014   69.39      1   
remove    -          -          -          -          -          -          12.11      1   
write     995.50     100663296  4.00       0.000026   98.75      0.000012   98.75      2   
read      1416.85    100663296  4.00       0.000014   69.38      0.000013   69.38      2   
remove    -          -          -          -          -          -          12.09      2   
write     1002.55    100663296  4.00       0.000027   98.05      0.000011   98.05      3   
read      1416.23    100663296  4.00       0.000014   69.41      0.000013   69.41      3   
remove    -          -          -          -          -          -          12.12      3   
write     999.25     100663296  4.00       0.000026   98.38      0.000012   98.38      4   
read      1408.93    100663296  4.00       0.000015   69.77      0.000014   69.77      4   
remove    -          -          -          -          -          -          12.11      4   

Max Write: 1002.55 MiB/sec (1051.25 MB/sec)
Max Read:  1417.42 MiB/sec (1486.27 MB/sec)

Test 1 started: Mon Feb 19 14:13:23 2018
Summary:
	api                = POSIX
	test filename      = /mnt/tmpfs/ior-bench.data
	access             = file-per-process
	ordering in a file = random offsets
	ordering inter file= random task offsets >= 1, seed=0
	clients            = 48 (48 per node)
	repetitions        = 5
	xfersize           = 4096 bytes
	blocksize          = 2 GiB
	aggregate filesize = 96 GiB

access    bw(MiB/s)  block(KiB) xfer(KiB)  open(s)    wr/rd(s)   close(s)   total(s)   iter
------    ---------  ---------- ---------  --------   --------   --------   --------   ----
write     25012      2097152    4.00       0.012435   3.93       0.561413   3.93       0   
read      22660      2097152    4.00       0.011624   4.34       2.70       4.34       0   
remove    -          -          -          -          -          -          0.360676   0   
write     25110      2097152    4.00       0.012490   3.91       0.516111   3.91       1   
read      22369      2097152    4.00       0.011774   4.39       2.81       4.39       1   
remove    -          -          -          -          -          -          0.366119   1   
write     25051      2097152    4.00       0.012591   3.92       0.696023   3.92       2   
read      22330      2097152    4.00       0.011762   4.40       2.81       4.40       2   
remove    -          -          -          -          -          -          0.363624   2   
write     25155      2097152    4.00       0.012653   3.91       0.596007   3.91       3   
read      22388      2097152    4.00       0.011746   4.39       2.86       4.39       3   
remove    -          -          -          -          -          -          0.365270   3   
write     25059      2097152    4.00       0.012512   3.92       0.550579   3.92       4   
read      22370      2097152    4.00       0.011739   4.39       2.81       4.39       4   
remove    -          -          -          -          -          -          0.363023   4   

Max Write: 25154.87 MiB/sec (26376.80 MB/sec)
Max Read:  22659.51 MiB/sec (23760.22 MB/sec)

Test 2 started: Mon Feb 19 14:14:07 2018
Summary:
	api                = POSIX
	test filename      = /mnt/tmpfs/ior-bench.data
	access             = file-per-process
	ordering in a file = sequential offsets
	ordering inter file= random task offsets >= 1, seed=0
	clients            = 48 (48 per node)
	repetitions        = 5
	xfersize           = 512 MiB
	blocksize          = 2 GiB
	aggregate filesize = 96 GiB

access    bw(MiB/s)  block(KiB) xfer(KiB)  open(s)    wr/rd(s)   close(s)   total(s)   iter
------    ---------  ---------- ---------  --------   --------   --------   --------   ----
write     22386      2097152    524288     0.012629   4.39       0.623432   4.39       0   
…